refactor test plan api
