--- docs/api_reference/test_plan_templates.rst
@@ -1,16 +1,0 @@
-.. _api_tag_page:
-
-nisystemlink.clients.test_plan.test_plan_templates
-======================
-
-.. autoclass:: nisystemlink.clients.test_plan.test_plan_templates.TestPlanTemplateClient
-   :exclude-members: __init__
-
-   .. automethod:: __init__
-   .. automethod:: create_test_plan_templates
-   .. automethod:: query_test_plan_templates
-   .. automethod:: delete_test_plan_templates
-
-.. automodule:: nisystemlink.clients.test_plan.test_plan_templates.models
-   :members:
-   :imported-members:
