+++ docs/getting_started.rst
@@ -343,32 +343,12 @@ Examples
 
 Create, query, get, update, schedule and delete TestPlans
 
-.. literalinclude:: ../examples/test_plan/test_plan/test_plans.py
+.. literalinclude:: ../examples/test_plan/test_plans.py
    :language: python
    :linenos:
 
-TestPlanTemplate API
--------
-
-Overview
-~~~~~~~~
-
-The :class:`.TestPlanTemplateClient` class is the primary entry point of the TestPlanTemplate API.
-
-When constructing a :class:`.TestPlanTemplateClient`, you can pass an
-:class:`.HttpConfiguration` (like one retrieved from the
-:class:`.HttpConfigurationManager`), or let :class:`.TestPlanTemplateClient` use the
-default connection. The default connection depends on your environment.
-
-With a :class:`.TestPlanTemplateClient` object, you can:
-
-* Create, query, and delete TestPlanTemplate
-
-Examples
-~~~~~~~~
-
 Create, query and delete test plan templates.
 
-.. literalinclude:: ../examples/test_plan/test_plan_templates/test_plan_templates.py
+.. literalinclude:: ../examples/test_plan/test_plan_templates.py
    :language: python
    :linenos:
