+++ nisystemlink/clients/test_plan/__init__.py
@@ -1,1 +1,3 @@
+from nisystemlink.clients.test_plan._test_plan_client import TestPlanClient
+
 # flake8: noqa
