--- nisystemlink/clients/test_plan/test_plan/__init__.py
@@ -1,3 +1,0 @@
-from ._test_plan_client import TestPlanClient
-
-# flake8: noqa
