--- nisystemlink/clients/test_plan/test_plan/models/__init__.py
@@ -1,23 +1,0 @@
-from ._create_test_plans_request import CreateTestPlansRequest
-from ._create_test_plan_request import CreateTestPlanRequest
-from ._create_test_plans_response import CreateTestPlansResponse
-from ._schedule_test_plan_request import ScheduleTestPlanRequest
-from ._schedule_test_plans_response import ScheduleTestPlansResponse
-from ._delete_test_plans_request import DeleteTestPlansRequest
-from ._execution_event import ExecutionEvent
-from ._query_test_plans_request import (
-    QueryTestPlansRequest,
-    TestPlanField,
-    _QueryTestPlansRequest,
-)
-from ._schedule_test_plans_request import ScheduleTestPlansRequest
-from ._state import State
-from ._test_plan import TestPlan
-from ._update_test_plan_request import UpdateTestPlanRequest
-from ._update_test_plans_request import UpdateTestPlansRequest
-from ._update_test_plans_response import UpdateTestPlansResponse
-from ._workflow_definition import WorkflowDefinition
-from ._order_by import OrderBy
-from ._paged_test_plans import PagedTestPlans
-
-# flake8: noqa
