--- nisystemlink/clients/test_plan/test_plan_templates/__init__.py
@@ -1,5 +1,0 @@
-from nisystemlink.clients.test_plan.test_plan_templates._test_plan_templates_client import (
-    TestPlanTemplateClient,
-)
-
-# flake8: noqa
