--- tests/integration/test_plan/__init__.py
@@ -1,1 +1,0 @@
-# flake8: noqa
