--- tests/integration/test_plan/test_plan_templates/__init__.py
@@ -1,1 +1,0 @@
-# flake8: noqa
